pyproject.toml
[tool.poetry]
name = "ga4-analytics-api"
version = "0.1.0"
description = "A project to access GA4 data via the Google Analytics Data API and export it to CSV and also to query google analytics websites. "
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
google-analytics-data = "^0.18.11"
pandas = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
